Add a full-submission cypress test and re-use login code
